Create new lesson for SEIr-129
Removed unneccessary files to start with a clean slate.  Using the
original repo to preserve history.  Will pattern this lesson more on
the JSD and Boston campus lessons which provide a better introduction
to React by focusing on the use of functional stateless components and
only introducing class syntax for state.